Fix: (prism-whitespace-mode) Depth detection inside lists
[Branch name means "indentation-delimited-inside-character-delimited".]

This is a very simple change that apparently fixes the calculation of
depth of indentation-delimited lists inside character-delimited
lists (e.g. in shell scripts, the indentation of if/else
clauses (indentation-delimited) inside a function ({}-delimited)).

It seems to work, but it also seems a bit too easy, so I'm going to
test it before merging.